requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.3
b5==1.4.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bs4==0.0.1
celery==4.4.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.6.2
flower==0.9.7
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
kombu==4.6.11
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
pymemcache==4.0.0
pypng==0.20220715.0
pyTelegramBotAPI==4.13.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==2.3.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==1.3.0
wcwidth==0.2.6
